+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   * get\_register becomes get\_register\_data and now returns an array of RegisterData objects
   * get\_register\_structure now returns an array of RegisterField objects
   * Add from\_api method to ApiAccess
+  * Add from\_api method to DueBadges
 
 ## Version 0.0.11